chore: Preparation for first official public release (#80)
