+++ .gitignore
@@ -1,5 +1,5 @@
 # Build output
-dist/
+types/
 
 # Build cache
 .tsup
