+++ jsr.json
@@ -1,10 +1,8 @@
 {
 	"$schema": "https://jsr.io/schema/config-file.v1.json",
 	"name": "@xeho91/svelte-ast-print",
-	"version": "0.2.0-next.0",
-	"exports": {
-		".": "./src/mod.js"
-	},
+	"version": "0.1.0",
+	"exports": "./src/mod.js",
 	"publish": {
 		"include": ["src/", "README.md"]
 	}
