+++ tsconfig.json
@@ -1,8 +1,8 @@
 {
 	"$schema": "https://json.schemastore.org/tsconfig",
-	"extends": ["@total-typescript/tsconfig/bundler/dom/library"],
+	"extends": "@total-typescript/tsconfig/tsc/dom/library",
 	"compilerOptions": {
 		"customConditions": ["development"]
 	},
-	"include": ["src/", "tests/"]
+	"include": ["src/"]
 }
